requirements.txt
# Core Machine Learning
tensorflow>=2.10.0
torch>=1.12.0
torchvision>=0.13.0
transformers>=4.20.0
scikit-learn>=1.1.0
numpy>=1.21.0
pandas>=1.4.0

# Computer Vision
pillow>=9.0.0
opencv-python>=4.6.0
albumentations>=1.2.0

# Vector Database
qdrant-client>=1.1.0

# Data Processing
pyyaml>=6.0
h5py>=3.7.0
scipy>=1.8.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0

# Web App
streamlit>=1.12.0
flask>=2.2.0

# Development Tools
jupyter>=1.0.0
jupyterlab>=3.4.0
ipywidgets>=7.7.0

# Testing
pytest>=7.1.0
pytest-cov>=3.0.0

# Utilities
tqdm>=4.64.0
requests>=2.28.0
pathlib2>=2.3.0

# Optional: For CUDA support (uncomment if using GPU)
# tensorflow-gpu>=2.10.0

# Optional: For advanced data augmentation
# imgaug>=0.4.0

# Optional: For model versioning and experiment tracking
# mlflow>=1.28.0
# wandb>=0.13.0

# Optional: For deployment
# gunicorn>=20.1.0
# docker>=6.0.0
